move files out of pages directory
