+++ src/ogv-dashboard/chart-configs/doughnutOptions.ts
