+++ src/ogv-dashboard/chart-configs/index.ts
