+++ src/ogv-dashboard/chart-configs/priceLineOptions.ts
