+++ src/ogv-dashboard/chart-configs/stakeLineOptions.ts
