+++ src/ogv-dashboard/components/Section.tsx
