+++ src/ogv-dashboard/sections/Heading.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Image from "next/image";
 import { Section } from "../components";
-import { assetRootPath } from "../../../src/utils/image";
+import { assetRootPath } from "../../utils/image";
 import { Button } from "@originprotocol/origin-storybook";
 
 interface HeadingProps {
