+++ src/ogv-dashboard/sections/OgvPriceStats.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useRef } from "react";
 import Image from "next/image";
-import shortenAddress from "../../../src/utils/shortenAddress";
+import shortenAddress from "../../utils/shortenAddress";
 import { Section } from "../components";
-import { useOutOfBoundsClick } from "../../../src/hooks";
-import { assetRootPath } from "../../../src/utils/image";
+import { useOutOfBoundsClick } from "../../hooks";
+import { assetRootPath } from "../../utils/image";
 import { BigNumber, utils } from "ethers";
 const { commify, formatEther } = utils;
