+++ src/ogv-dashboard/types/DashProps.ts
