+++ src/ogv-dashboard/types/GetStorageAtResponse.ts
