+++ src/ogv-dashboard/types/Link.ts
