+++ src/ogv-dashboard/utils/get24HChartData.ts
