+++ src/ogv-dashboard/utils/getStakingChartData.ts
